Support tidb chunk's data format and columnar execution in Spark  (#1220)

+++ core/src/test/scala/org/apache/spark/sql/catalyst/catalog/CatalogTestSuite.scala
@@ -72,7 +72,7 @@ class CatalogTestSuite extends BaseTiSparkTest {
         List("tp_bigint", "bigint", "true", null),
         List("tp_decimal", "decimal(11,0)", "true", null),
         List("tp_double", "double", "true", null),
-        List("tp_float", "double", "true", null),
+        List("tp_float", "float", "true", null),
         List("tp_int", "bigint", "true", null),
         List("tp_mediumint", "bigint", "true", null),
         List("tp_real", "double", "true", null),
